putting run, run_optmized into a single file run.py
