+++ classic_ml/LinearRegression/run.py
@@ -8,16 +8,8 @@ from utils.benchmark import run_model
 from classic_ml.tabular_dataset import TabularDataset
 from utils.misc import print_goodbye_message_and_die, download_ampere_imagenet
 
-from sklearn.linear_model import  LinearRegression
-
-# from sklearn.linear_model import  LinearRegression as LinearRegressionBase
-# class LinearRegression(LinearRegressionBase):
-#     #overwriting decision func from original scikit learn and in this it skips checking and validation
-#     def _decision_function(self, X):
-#         return safe_sparse_dot(X, self.coef_.T, dense_output=True) + self.intercept_
-
 def parse_args():
-    parser = argparse.ArgumentParser(description="Run DecisionTreeClassifier model.")
+    parser = argparse.ArgumentParser(description="Run LinearRegression model.")
     parser.add_argument("-m", "--model_path",
                         type=str,
                         help="path to the model")
@@ -40,10 +32,12 @@ def parse_args():
     parser.add_argument("--data_path",
                         type=str,
                         help="path to csv file containing input data")
+    parser.add_argument("--optimized",
+                        type=bool,
+                        default=False,
+                        help="whether optimized LinearRegression is used or not")
     args = parser.parse_args()
-   
-    if args.framework != "sklearn" and args.model_path is None:
-        parser.error(f"You need to specify the model path when using {args.framework} framework.")
+
     return args
 
 def run_ort_fp(model_path, batch_size, num_runs, timeout, data_path):
@@ -61,7 +55,21 @@ def run_ort_fp(model_path, batch_size, num_runs, timeout, data_path):
     
     return run_model(run_single_pass, runner, dataset, batch_size, num_runs, timeout)
 
-def run_sklearn_fp(model_path, batch_size, num_runs, timeout, data_path):
+def run_sklearn_fp(model_path, batch_size, num_runs, timeout, data_path, optimized):
+    global LinearRegression
+    if optimized:
+        print('Running in Optimized Mode')
+        from sklearn.linear_model import  LinearRegression as LinearRegressionBase
+        from sklearn.linear_model._base import safe_sparse_dot
+        class LinearRegression(LinearRegressionBase):
+            #overwriting decision func from original scikit learn and in this it skips checking and validation
+            def _decision_function(self, X):
+                return safe_sparse_dot(X, self.coef_.T, dense_output=True) + self.intercept_
+
+    else:
+        from sklearn.linear_model import  LinearRegression
+
+    print(LinearRegression)
     from utils.sklearn import SklearnRunner
 
     def run_single_pass(sklearn_runner, dataset):
@@ -78,8 +86,8 @@ def run_sklearn_fp(model_path, batch_size, num_runs, timeout, data_path):
 def run_ort_fp32(model_path, batch_size, num_runs, timeout, data_path, **kwargs):
     return run_ort_fp(model_path, batch_size, num_runs, timeout, data_path)
 
-def run_sklearn_fp32(model_path, batch_size, num_runs, timeout, data_path, **kwargs):
-    return run_sklearn_fp(model_path, batch_size, num_runs, timeout, data_path)
+def run_sklearn_fp32(model_path, batch_size, num_runs, timeout, data_path, optimized, **kwargs):
+    return run_sklearn_fp(model_path, batch_size, num_runs, timeout, data_path, optimized)
